backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
sentence-transformers==2.2.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
PyGithub==2.1.1
python-multipart==0.0.6